+++ py42/_internal/clients/storage/security.py
@@ -5,7 +5,7 @@ from py42._internal.base_classes import BaseStorageClient
 
 class StorageSecurityClient(BaseStorageClient):
 
-    def get_security_detection_events(self, user_uid=None, plan_uid=None, cursor=None, include_files=None,
+    def _get_security_detection_events(self, user_uid=None, plan_uid=None, cursor=None, include_files=None,
                                       event_types=None, min_timestamp=None, max_timestamp=None, summarize=None,
                                       **kwargs):
         uri = "/api/SecurityDetectionEvent"
@@ -26,25 +26,23 @@ class StorageSecurityClient(BaseStorageClient):
 
     def get_security_detection_events_for_plan(self, plan_uid, cursor=None, include_files=None, event_types=None,
                                                min_timestamp=None, max_timestamp=None, **kwargs):
-        return self.get_security_detection_events(plan_uid=plan_uid, cursor=cursor,
-                                                  include_files=include_files,
-                                                  event_types=event_types,
-                                                  min_timestamp=min_timestamp,
-                                                  max_timestamp=max_timestamp, **kwargs)
+        return self._get_security_detection_events(plan_uid=plan_uid, cursor=cursor,
+                                                   include_files=include_files,
+                                                   event_types=event_types,
+                                                   min_timestamp=min_timestamp,
+                                                   max_timestamp=max_timestamp, **kwargs)
 
     def get_security_detection_events_for_user(self, user_uid, cursor=None, include_files=None, event_types=None,
                                                min_timestamp=None, max_timestamp=None, **kwargs):
-        return self.get_security_detection_events(user_uid=user_uid, cursor=cursor,
-                                                  include_files=include_files,
-                                                  event_types=event_types,
-                                                  min_timestamp=min_timestamp,
-                                                  max_timestamp=max_timestamp, **kwargs)
+        return self._get_security_detection_events(user_uid=user_uid, cursor=cursor,
+                                                   include_files=include_files,
+                                                   event_types=event_types,
+                                                   min_timestamp=min_timestamp,
+                                                   max_timestamp=max_timestamp, **kwargs)
 
-    def get_security_detection_event_summary(self, user_uid, cursor=None, include_files=None, event_types=None,
-                                             min_timestamp=None, max_timestamp=None, **kwargs):
-        return self.get_security_detection_events(user_uid=user_uid, cursor=cursor,
-                                                  include_files=include_files,
-                                                  event_types=event_types,
-                                                  min_timestamp=min_timestamp,
-                                                  max_timestamp=max_timestamp,
-                                                  summarize=True, **kwargs)
+    def get_security_detection_event_summary(self, user_uid, cursor=None, min_timestamp=None, max_timestamp=None,
+                                             **kwargs):
+        return self._get_security_detection_events(user_uid=user_uid, cursor=cursor,
+                                                   min_timestamp=min_timestamp,
+                                                   max_timestamp=max_timestamp,
+                                                   summarize=True, **kwargs)